fixed compilation and reduced warnings
git-svn-id: file:///home/karl/SVN2/amilab.svn/code/trunk@902 52b0d089-85ac-4fb5-acf5-4e3172ea7aaf

+++ src/Graphic/Common/dwControlledCurve.cpp
@@ -15,7 +15,7 @@
 void dwControlledCurve::ComputeCurve()
 {
   dwCurve::ptr curve(new dwCurve());
-  for(int i=0; i<_controlpoints->size(); i++)
+  for(int i=0; i<(int)_controlpoints->size(); i++)
   {
     curve->AddPoint(dwPoint2D((*_controlpoints)[i].GetX(),(*_controlpoints)[i].GetY()));
   }
